:shirt: refactor: code formatting
