+++ tinygo/v0/net/syscall.go
@@ -1,7 +1,6 @@
 package net
 
 import (
-	"fmt"
 	"syscall"
 )
 
@@ -9,7 +8,7 @@ func syscallControlFd(rawConn syscall.RawConn, f func(fd uintptr) error) (err er
 	if controlErr := rawConn.Control(func(fd uintptr) {
 		err = f(fd)
 	}); controlErr != nil {
-		panic(fmt.Sprintf("controlErr = %v", controlErr))
+		// panic(fmt.Sprintf("controlErr = %v", controlErr))
 		return controlErr
 	}
 	return err
