+++ tinygo/v0/net/syscall_unix.go
@@ -7,3 +7,5 @@ import "syscall"
 func syscallSetNonblock(fd uintptr, nonblocking bool) (err error) {
 	return syscall.SetNonblock(int(fd), nonblocking)
 }
+
+type syscallFd = int
